==> 2. Scraped Data merging.ipynb <==
import pandas as pd

data_jan_to_mar=pd.read_csv(r"C:\Users\WELCOME\Downloads\data\data_31_dec_to_25_mar.csv")
data_jan_to_mar.shape

data_mar_to_may=pd.read_csv(r"C:\Users\WELCOME\Downloads\data\data_26Mar_to_29May.csv")
data_mar_to_may.shape

data_may_to_july=pd.read_csv(r"C:\Users\WELCOME\Downloads\data\data_30_May_to_26_july.csv")
data_may_to_july.shape

data_july=pd.read_csv(r"C:\Users\WELCOME\Downloads\data\data_27july_to_30july .csv")
data_july.shape

data_july_to_oct=pd.read_csv(r"C:\Users\WELCOME\Downloads\data\data_31_july_to_26_oct .csv")
data_july_to_oct.shape

data_oct_to_dec=pd.read_csv(r"C:\Users\WELCOME\Downloads\data\data_27_oct_to_14_dec.csv")
data_oct_to_dec.shape

data_m1=pd.concat([data_jan_to_mar,data_mar_to_may],axis=0)

data_m2=pd.concat([data_m1,data_may_to_july],axis=0)

data_m3=pd.concat([data_m2,data_july],axis=0)

data_m4=pd.concat([data_m3,data_july_to_oct],axis=0)

data_m5=pd.concat([data_m4,data_oct_to_dec],axis=0)

data_m5.shape

data_m5.to_csv(r"C:\Users\WELCOME\Downloads\data\final_dataset.csv",index=False)
