==> 3. data_preprocessing_final.ipynb <==
#nltk.download('all')

import pandas as pd
import nltk
#from nrclex import NRCLex
from nltk import word_tokenize
from nltk.corpus import stopwords
from nltk.stem.porter import PorterStemmer
from nltk.stem import WordNetLemmatizer 
from nltk.sentiment.vader import SentimentIntensityAnalyzer
import string
import os, re, csv, json, sys, string
import numpy as np
import pandas as pd
from collections import defaultdict, Counter

import gzip

from tqdm import tqdm

import pickle as pkl
from argparse import ArgumentParser
import logging
import pandas as pd
from nltk.stem.wordnet import WordNetLemmatizer
from nltk.corpus import wordnet

#still checking extension

strings_punc=string.punctuation + "\n" +"\t"

def get_wordnet_pos(tag):

    if tag.startswith('J'):
        return wordnet.ADJ
    elif tag.startswith('V'):
        return wordnet.VERB
    elif tag.startswith('N'):
        return wordnet.NOUN
    elif tag.startswith('R'):
        return wordnet.ADV
    else:
        return wordnet.NOUN

def text_preprocessing(text):
    mapped=[]
    lemmatized_output=[]
    text_p = "".join([char for char in text if char not in strings_punc])
    text=text_p.strip()
    text=text.replace('“', '').replace('”', '')
    text=text.replace('’', '').replace('’', '')
    text=text.replace('', '')
    new_text = ''.join((x for x in text if not x.isdigit()))
    text=new_text.lower()
    #print(new_text)
    words = word_tokenize(new_text)
    stop_words = stopwords.words('english')
    filtered_words = [word for word in words if word not in stop_words]
    #print(type(filtered_words))
    tagged = nltk.pos_tag(filtered_words)
    
    #print(tagged[0][1])
    #print(type(tagged[0][1]))
    lemmatizer = WordNetLemmatizer()

    for i in range(len(filtered_words)):
        m=get_wordnet_pos(tagged[i][1])
        #print(m)
        mapped.append(m)
        
        lemmatized_output.append(lemmatizer.lemmatize(filtered_words[i],mapped[i]))
    
    #print(type(mapped[0]))
    return(lemmatized_output)
    
    
    
    
    
    

#text_preprocessing(data['news_text'][0])


data=pd.read_csv(r"C:\Users\WELCOME\Downloads\data\final_dataset.csv",index_col=False)
data.news_text=data.news_text.astype(str)
    

data['news_text']=data['news_text'].apply(lambda x: text_preprocessing(x))

#example of preprocessed text

data['news_text'][1]

data.rename(columns = {'news_text':'text'}, inplace = True)

data['speaker']='UK'

data.head()

data.shape

def isfloat(value):
  try:
    float(value)
    return True
  except ValueError:
    return False

#deleting rows with NaNs

data['check_val']=data.text.apply(lambda x:isfloat(x[0]) )

data_filtered=data[data.check_val==False]
data_filtered.shape

#deleted 18 rows

data_filtered.to_csv(r"C:\Users\WELCOME\Downloads\data\preprocessed_data_final.csv",index=False)
