==> 8. Reading Covid Data.ipynb <==
import pandas as pd

data=pd.read_csv(r"C:\Users\WELCOME\Downloads\data\data_2022-Aug-24.csv")

data

data.columns

new_data=data[data.columns[3:10]]

new_data.head()

type(new_data.date[0])

new_data.date=pd.to_datetime(new_data.date)

new_data.head()

new_data_2021=new_data[(new_data.date>=pd.to_datetime("2021-01-01")) & (new_data.date<=pd.to_datetime("2021-12-31"))]

new_data_2021.shape

new_data_2021.to_csv(r"C:\Users\WELCOME\Downloads\data\covid_data_2021",index=False)
