==> 9. Post correlation test Analysis.ipynb <==
import pandas as pd
from textblob import TextBlob, Word, Blobber
import matplotlib.pyplot as plt
from wordcloud import WordCloud, STOPWORDS, ImageColorGenerator
import numpy as np
import re, string, timeit
import wordninja
from sklearn.feature_extraction.text import CountVectorizer
from plotly.offline import download_plotlyjs, init_notebook_mode, plot, iplot
import cufflinks as cf
cf.go_offline()
cf.set_config_file(offline=False, world_readable=True)

emo_data=pd.read_csv(r"C:\Users\WELCOME\Downloads\data\text_emo_data.csv",index_col=False)

def get_top_n_words(corpus, n=None):
    vec = CountVectorizer(stop_words='english').fit(corpus)
    bag_of_words = vec.transform(corpus)
    sum_words = bag_of_words.sum(axis=0) 
    words_freq = [(word, sum_words[0, idx]) for word, idx in vec.vocabulary_.items()]
    words_freq =sorted(words_freq, key = lambda x: x[1], reverse=True)
    return words_freq[:n]

disgust_text=emo_data[emo_data.overall_emo=='avgLexVal_disgust'].text.str.cat(sep=' ')

joy_text=emo_data[emo_data.overall_emo=='avgLexVal_joy'].text.str.cat(sep=' ')

print("Frequent Words disgust text")
common_words = get_top_n_words(emo_data[emo_data.overall_emo=='avgLexVal_disgust'].text, 20)
total=0
for word, freq in common_words:

    print(word, freq)
    
    total=total+freq
for word, freq in common_words:
    perc=(freq/len(disgust_text))*100
    print(word,str(perc)+"%")
df1 = pd.DataFrame(common_words, columns = ['ReviewText' , 'count'])
df1.groupby('ReviewText').sum()['count'].sort_values(ascending=False).iplot(
    kind='bar', yTitle='Count', linecolor='black', title='Top 20 words  from disgust corpus')

print("Frequent Words joy text")
common_words = get_top_n_words(emo_data[emo_data.overall_emo=='avgLexVal_joy'].text, 20)
total=0
for word, freq in common_words:

    print(word, freq)
    
    total=total+freq
for word, freq in common_words:
    perc=(freq/len(joy_text))*100
    print(word,str(perc)+"%")
df1 = pd.DataFrame(common_words, columns = ['ReviewText' , 'count'])
df1.groupby('ReviewText').sum()['count'].sort_values(ascending=False).iplot(
    kind='bar', yTitle='Count', linecolor='black', title='Top 20 words  from joy corpus')

## Bigrams

def get_top_n_bigram(corpus, n=None):
    vec = CountVectorizer(ngram_range=(2, 2),stop_words='english').fit(corpus)
    bag_of_words = vec.transform(corpus)
    sum_words = bag_of_words.sum(axis=0) 
    words_freq = [(word, sum_words[0, idx]) for word, idx in vec.vocabulary_.items()]
    words_freq =sorted(words_freq, key = lambda x: x[1], reverse=True)
    return words_freq[:n]

print("Frequent Bigrams for disgust corpus")
common_words = get_top_n_bigram(emo_data[emo_data.overall_emo=='avgLexVal_disgust'].text, 20)
total=0
for word, freq in common_words:

    print(word, freq)
    
    total=total+freq
for word, freq in common_words:
    perc=(freq/len(disgust_text))*100
    print(word,str(perc)+"%")
df3 = pd.DataFrame(common_words, columns = ['ReviewText' , 'count'])
df3.groupby('ReviewText').sum()['count'].sort_values(ascending=False).iplot(
    kind='bar', yTitle='Count', linecolor='black', title='Top 20 bigrams for disgust corpus')

print("Frequent Bigrams for joy corpus")
common_words = get_top_n_bigram(emo_data[emo_data.overall_emo=='avgLexVal_joy'].text, 20)
total=0
for word, freq in common_words:

    print(word, freq)
    
    total=total+freq
for word, freq in common_words:
    perc=(freq/len(joy_text))*100
    print(word,str(perc)+"%")
df3 = pd.DataFrame(common_words, columns = ['ReviewText' , 'count'])
df3.groupby('ReviewText').sum()['count'].sort_values(ascending=False).iplot(
    kind='bar', yTitle='Count', linecolor='black', title='Top 20 bigrams for joy corpus')
